// apps/api/src/modules/notifications/transactional/notification-event-type.ts
export enum NotificationEventType {
	ORDER_PLACED_CHECKOUT = "ORDER_PLACED_CHECKOUT",
	ORDER_PLACED_BUYNOW = "ORDER_PLACED_BUYNOW",
	ORDER_PAYMENT_RECEIVED_CHECKOUT = "ORDER_PAYMENT_RECEIVED_CHECKOUT",
	ORDER_PAYMENT_RECEIVED_BUYNOW = "ORDER_PAYMENT_RECEIVED_BUYNOW",
	ORDER_PAYMENT_FAILED = "ORDER_PAYMENT_FAILED",
	ORDER_CANCELLED = "ORDER_CANCELLED",
	COURSE_ACCESS_GRANTED = "COURSE_ACCESS_GRANTED",
	TRAINING_ENROLLMENT_CONFIRMED = "TRAINING_ENROLLMENT_CONFIRMED",
	BOOKLET_ACCESS_GRANTED = "BOOKLET_ACCESS_GRANTED",
	ORDER_REFUND_REQUESTED = "ORDER_REFUND_REQUESTED",
	ORDER_REFUND_APPROVED = "ORDER_REFUND_APPROVED",
	ORDER_REFUND_REJECTED = "ORDER_REFUND_REJECTED",
}